+++ lib/pod/command/trunk.rb
@@ -70,16 +70,32 @@ module Pod
         raise Informative, error
       end
 
-      def print_messages(data_url, messages)
-        # Using UI.labeled here is dangerous, as it wraps the URL and indents
-        # it, which breaks the URL when you try to copy-paste it.
-        UI.puts "  - Data URL: #{data_url}"
+      def print_messages(data_url, messages, spec = nil, action = nil)
+        if verbose? || spec.nil?
+          # Using UI.labeled here is dangerous, as it wraps the URL and indents
+          # it, which breaks the URL when you try to copy-paste it.
+          UI.puts "  - Data URL: #{data_url}"
 
-        messages = messages.map do |entry|
-          at, message = entry.to_a.flatten
-          "#{formatted_time(at)}: #{message}"
+          server_logs = messages.map do |entry|
+            at, message = entry.to_a.flatten
+            "#{formatted_time(at)}: #{message}"
+          end
+          UI.labeled 'Log messages', server_logs
+        else
+          separator = '-' * 80
+          UI.puts
+          UI.puts separator
+          UI.puts " 🎉  Congrats"
+          UI.puts
+          UI.puts " 🚀  #{spec.name} (#{spec.version}) successfully #{action}"
+          unless messages.empty?
+            at = messages.first.to_a.flatten.first
+            UI.puts " 📅  #{formatted_time(at)}"
+          end
+          UI.puts " 🌎  https://cocoapods.org/pods/#{spec.name}"
+          UI.puts " 👍  Tell your friends!"
+          UI.puts separator
         end
-        UI.labeled 'Log messages', messages
       end
 
       def json(response)