+++ lib/cocoapods_trunk.rb
@@ -1,3 +1,3 @@
 module CocoaPodsTrunk
-  VERSION = '1.1.0.beta.1'.freeze
+  VERSION = '1.1.0'.freeze
 end