+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## Master
+
+##### Enhancements
+
+* None.  
+
+##### Bug Fixes
+
+* None.  
+
+
 ## 1.1.0.beta.1 (2016-10-10)
 
 ##### Enhancements