+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Master
+## 1.1.0 (2016-10-19)
 
 ##### Enhancements